i need to somehow manage the complexity XD

+++ src/cube.h
@@ -42,7 +42,7 @@ static const Vertex cube_vertices[] = {
   {{-0.5f,  0.5f, -0.5f}, {1.0f, 0.0f, 1.0f}}, // 23
 };
 
-static const GLushort cube_indices[] = {
+static const GLuint cube_indices[] = {
   // front face
   0, 2, 3,
   2, 0, 1,
